requirements.txt
numpy
matplotlib
pandas
scikit-learn
torch==1.13.0
python-dotenv
requests
pandas
asyncio
aiohttp